fixed logout-login bug
