+++ Waffle-Project/Waffle-Clone/Controller/LoginViewController.swift
@@ -10,37 +10,34 @@ import UIKit
 
 class LoginViewController: UIViewController {
     @IBOutlet weak var loginBtn: UIButton!
-    let authenticationManager = AuthenticationManager()
-   
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        authenticateUser()	
-    }
-    
+    let loginManager = LoginManager()
+    let authenticationAlert = Alert.showBasicAlert(with: "Error", message: "Authentication failed.")
+
+
     override func viewDidLoad() {
         super.viewDidLoad()
-        // If user has never logged in
-        if AuthenticationManager.accessToken == nil {
-            let alert = Alert.showBasicAlert(with: "Error", message: "Authentication failed.")
-            self.present(alert, animated: true)
-        }
+        authenticateUser()
     }
     
     @IBAction func loginBtnPressed(_ sender: UIButton) {
-        if let authorizationUrl = authenticationManager.buildLoginURL(with: Scopes.allCases) {
-            UIApplication.shared.open(authorizationUrl)
+        loginManager.loginUser() { (success) in
+            if !success {
+                self.present(self.authenticationAlert, animated: true)
+            }
         }
     }
     
     /// If keychain contains a valid access token the user is sent to Repository View Controller. Otherwise, user is shown an error and sent to Login View Controller.
     private func authenticateUser() {
-        authenticationManager.hasValidToken() { (success) in
+        loginManager.hasValidToken() { (success) in
             if success {
-                if let vc = self.storyboard!.instantiateViewController(withIdentifier: "RepoViewController") as? RepositoryViewController {
-                    DispatchQueue.main.async {
-                        self.navigationController?.pushViewController(vc, animated: true)
+                DispatchQueue.main.async {
+                    if let vc = self.storyboard!.instantiateViewController(withIdentifier: "RepoViewController") as? RepositoryViewController {
+                            self.navigationController?.pushViewController(vc, animated: true)
                     }
                 }
+            } else {
+                self.present(self.authenticationAlert, animated: true)
             }
         }
     }
